refactor(onboarding): isolate install job output handling
